process upload when avatar null

+++ server/controllers/PatientController.js
@@ -38,15 +38,23 @@ module.exports = {
                             where: { id: id }
                         }).then((affectedCount) => {
                             let oldPath = path.resolve(__dirname, '..', '..', `uploads/${oldFileName}`);
-                            fs.unlink(oldPath, (error) => {
-                                if (error) {
-                                    return res.status(httpStatus.BAD_REQUEST).json(Util.error(error.message, httpStatus.BAD_REQUEST));
-                                }
+                            if(oldFileName){
+                                fs.unlink(oldPath, (error) => {
+                                    if (error) {
+                                        return res.status(httpStatus.BAD_REQUEST).json(Util.error(error.message, httpStatus.BAD_REQUEST));
+                                    }
+                                    Patient.findByPrimary(id).then(patient => {
+                                        return res.json(Util.success('Update patients success!', patient));
+                                    }).catch((error) => res.status(httpStatus.BAD_REQUEST).json(Util.error(error.message, httpStatus.BAD_REQUEST)));
+
+                                });
+                            }
+                            else{
                                 Patient.findByPrimary(id).then(patient => {
                                     return res.json(Util.success('Update patients success!', patient));
                                 }).catch((error) => res.status(httpStatus.BAD_REQUEST).json(Util.error(error.message, httpStatus.BAD_REQUEST)));
+                            }
 
-                            });
 
                         }).catch((error) => res.status(httpStatus.BAD_REQUEST).json(Util.error(error.message, httpStatus.BAD_REQUEST)));
                     });
